# Conflicts:
#	backend/src/main/java/pw/edu/pl/backend/repository/PlayRepository.java
#	backend/src/main/java/pw/edu/pl/backend/service/EventService.java

+++ backend/src/main/java/pw/edu/pl/backend/repository/EventRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pw.edu.pl.backend.entity.EventEn;
 
 @Repository
-public interface EventRepository extends JpaRepository<EventEn, Integer> {
+public interface EventRepository extends JpaRepository<EventEn, Long> {
 
 }
